Move Directives1 Interop into own file
